requirements.txt
fastapi
uvicorn
scikit-learn
numpy
python-docx
nltk
